resume-service/requirements.txt
fastapi
uvicorn
python-multipart
pika
pymongo
sentence-transformers
rank-bm25
pypdf
python-docx
scikit-learn
groq
python-dotenv
